Fix N+1 query in CreatorSearchOperation
The filtering in CreatorSearchOperation was issuing the same sql query
for every record in the results set, so up to 25 queries in total.

This change ensures the query only happens once. This is a tactical fix,
a better long term strategy would be to push this filtering to the
initial database query and avoid the need for additional queries at all.

I also deleted the test of the filtering in CreatorSearchOperationTest
since this filtering is delegated to CaseAccessService (where it is
now tested there).