CCD-3297: Changed location of smoke and functional dummy test results file
Changed location that DummyTest.xml is copied to for both the smoke and functional tests tasks.  The locations now match those expected by the stages of the pipeline that archive the smoke and functional test results.  These locations were recently corrected under DTSPO-6817.

The DummyTest.xml file is used because BEFTA tests are used for both smoke and functional tests.  These create test result files in a different location to that expected by the pipeline.