CCD-3396: Improve no validator found error logging
Added SLF4J logger to CaseDataValidator.  This is used to log an error message that includes the field name when no validator can be found for a field type.  This approach ensures that sensitive field names do not appear in the error message displayed in EXUI.  Following a review replaced the explicit slf4j setup with lombok @Slf4j annotation, removed try/catch around validatorToExecute() and added log statement to orElseThrow().

Added test for the changes to CaseDataValidatorTest class.  This required a change to TestConfiguration to add the Label base type.

Updated DefaultCaseDefinitionRepositoryIT.shouldGetBaseTypes() test to include Label base type.  Also corrected test so that it checked for presence of DateTime base type rather than Date base type twice.