Ccd 216 malformed case type id error response (#1979)
* CCD-216: Handle malformed case type ids

Created DataStoreHttpStatusRequestRejectedHandler class.  This handles RequestRejectedExceptions raised by the Spring StrictHttpFirewall that occur when the URL contains forbidden characters.  The handler converts the status code of the response from 500 to 403.  Created FirewallConfiguration class to register a bean for the new handler and make it available to Spring.  Created a new functional test - S-1032.2 - to test this scenario.

Created CaseTypeIdValidator class.  This is used to validate the case type id in requests that are accepted by StrictHttpFirewall.  Created ValidCaseTypeId annotation for CaseTypeIdValidator and updated CaseController to use it to validate case type ids.

Updated RestExceptionHandler to add a new handler for ConstraintViolationExceptions.  These exceptions are raised when the case type id fails validation.  The handler returns a response with the status code 400.  Updated HttpError class to add a new constructor which is used by the new handler.

Renamed S-1032 test to S-1032.1 and re-enabled it.  Changed expected response to match that returned by the new case type id validation.

* CCD-216: Updated case type id validation regular expression

Updated the regular expression used in the CaseTypeIdValidator class to validate the case type id.  Changed it to be the same as the REFERENCE_PATTERN regular expression used in ccd-definition-store-api CaseTypeEntityReferenceValidator class.

Co-authored-by: Rajaram <70951210+rajaram-nagarajan@users.noreply.github.com>
Co-authored-by: marianadpereira <71711509+marianadpereira@users.noreply.github.com>
Co-authored-by: mikebrownccd <104495891+mikebrownccd@users.noreply.github.com>
Co-authored-by: Dan Lysiak <50049163+danlysiak@users.noreply.github.com>