Fix flakey tests (#2564)
* Fix flakey shouldRetryOnErrorWithIgnoreWarningFalseAndDefaultRetryContext

This test relies on tight timings which causes it to regularly fail.
The final request had a delay of 490ms which is close to the 500ms
http timeout the test configures. The test scenario doesn't require a
delay for the final request.

* Fix flakey elasticsearch tests

Ensure we wait for indexing to finish before proceeding with tests.

* Fix flakey shouldReturn200WhenGetValidCaseDocuments test

The DocumentsOperationTest modifies the spring boot test context using
field level reflection which was breaking other tests.

@DirtiesContext ensures this modified context is not reused in
subsequent tests.

---------

Co-authored-by: Alex McAusland <a.m@user.com>